More clean up - Move RoutePattern parsing to separate file - Support PipeReader as body - Small renames
